pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "routeforge"
version = "0.1.0"
description = "CCNP-depth pedagogical router workbook with dual simulation abstractions"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "RouteForge" }]
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
routeforge = "routeforge.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra -q"
